Af branch

+++ backend/src/main/java/com/point/blood/users/UsersProfileDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -22,6 +24,6 @@ public class UsersProfileDTO {
     private String bloodGroup;
     private Character rhFactor;
     private LocalDate lastDonationDate;
-    private Double totalDonatedBlood;
+    private BigDecimal totalDonatedBlood;
 
 }
